move FuxNotification to shared components
