+++ web_ui/src/shared/components/fux-notification/fux-notification.module.scss
@@ -8,7 +8,7 @@
         fill: #89ad4e !important;
     }
 
-    margin-top: 5px !important;
+    margin-top: var(--spectrum-global-dimension-size-65) !important;
     min-width: var(--spectrum-global-dimension-size-4600) !important;
 }
