+++ web_ui/src/shared/components/header/credit-balance/credit-balance-button.component.tsx
@@ -17,10 +17,10 @@ import { useOrganizationIdentifier } from '../../../../hooks/use-organization-id
 import { usePrevious } from '../../../../hooks/use-previous/use-previous.hook';
 import { useProject } from '../../../../pages/project-details/providers/project-provider/project-provider.component';
 import { ONE_MINUTE } from '../../../utils';
+import { FuxNotification } from '../../fux-notification/fux-notification.component';
 import { useCheckPermission } from '../../has-permission/has-permission.component';
 import { OPERATION } from '../../has-permission/has-permission.interface';
 import { CreditsToConsume } from './credits-to-consume.component';
-import { FuxNotification } from './fux-notification/fux-notification.component';
 import { isBalanceLow } from './util';
 
 import classes from './credit-balance.module.scss';
